+++ designsafe/static/scripts/applications/controllers/application-systems-role.js
@@ -1,46 +1,52 @@
-export function applicationSystemsRoleCtrl(window, angular, $, _) {
-  'ngInject';
-  "use strict";
-  angular.module('designsafe').controller('ApplicationSystemsRoleCtrl',
-    ['$scope', '$rootScope', '$q', '$timeout', '$uibModal', '$translate', '$state', 'Apps', 'Django', function($scope, $rootScope, $q, $timeout, $uibModal, $translate, $state, Apps, Django) {
+import _ from 'underscore';
 
-        $scope.getSystemRoles = function(){
-          $scope.requesting = true;
-          var execSystem = $translate.instant('execution_default');
+export class ApplicationSystemsRoleCtrl {
+    constructor($translate, $state, Apps, Django) {
+        'ngInject';
+        this.$translate = $translate;
+        this.$state = $state;
+        this.Apps = Apps;
+        this.Django = Django;
+    }
 
-          Apps.getSystemRoles(execSystem)
-            .then(
-              function(response){
-                if (Django.user === 'ds_admin'){
-                  $scope.requesting = true;
-                  $state.go('applications-add-admin');
-                } else {
-                  _.each(response.data, function(role){
-                    if (role.username === Django.user){
-                      if (role.role === 'ADMIN' || role.role === 'PUBLISHER' || role.role === 'OWNER'){
-                        $state.go('applications-add');
-                      }
+    $onInit() {
+        this.getSystemRoles();
+    }
+
+    getSystemRoles() {
+        let self = this;
+        self.requesting = true;
+
+        if (this.Django.user === 'ds_admin') {
+            this.$state.go('applications-add-admin');
+        } else {
+            const execSystem = this.$translate.instant('execution_default');
+
+            this.Apps.getSystemRoles(execSystem)
+                .then(
+                    function(response) {
+                        _.each(response.data, function(role) {
+                            if (role.username === self.Django.user) {
+                                if (role.role === 'ADMIN' || role.role === 'PUBLISHER' || role.role === 'OWNER') {
+                                    self.$state.go('applications-add');
+                                }
+                            }
+                        });
+                        self.requesting = false;
+                    },
+                    function(response) {
+                        if (response.data) {
+                            if (response.data.message) {
+                                self.error = self.$translate.instant('error_app_system_roles') + response.data.message;
+                            } else {
+                                self.error = self.$translate.instant('error_app_system_roles') + response.data;
+                            }
+                        } else {
+                            self.error = self.$translate.instant('error_app_system_roles');
+                        }
+                        self.requesting = false;
                     }
-                  });
-                  $scope.requesting = false;
-                }
-              },
-              function(response){
-                if (response.data) {
-                  if (response.data.message){
-                    $scope.error = $translate.instant('error_app_system_roles') + response.data.message;
-                  } else {
-                    $scope.error = $translate.instant('error_app_system_roles') + response.data;
-                  }
-                } else {
-                  $scope.error = $translate.instant('error_app_system_roles');
-                }
-                $scope.requesting = false;
-              }
-            );
-      };
-
-      $scope.getSystemRoles();
-
-    }]);
-  }
+                );
+        }
+    }
+}