+++ designsafe/static/scripts/data-depot/controllers/main.js
@@ -1,3 +1,0 @@
-export function MainCtrl($scope, DataBrowserService) {
-    $scope.browser = DataBrowserService.state();
-  }