+++ designsafe/static/scripts/ng-designsafe/components/index.js
@@ -1,0 +1,5 @@
+import './modals';
+
+let designsafeComponents = angular.module('designsafe.components', ['designsafe.components.modals']);
+
+export default designsafeComponents;